pyproject.toml
[project]
name = "mlx-vision-agent"
version = "0.1.0"
description = "MLX Vision Agent using Qwen2.5-VL"
requires-python = ">=3.9"
dependencies = [
    "requests",
    "pillow",
    "mss",
    "python-dotenv",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]